refactor(RacingGame.java): RacingGame 클래스 로직 두개의 클래스로 분리
하나의 RacingGame 클래스에 너무 많은 로직이 들어가 코드의 가독성이
떨어지는 문제 해결을 위해 RacingGame클래스를 두개의 클래스로 분리.

BREAKING CHANGE: RacingGame클래스를 RacingGame과 Race 두개의 클래스로
분리. RacingGame에서는 레이싱 게임에 필요한 정보를 입력받고 Race에서는
RacingGame에서 입력받는 정보를 토대로 레이싱 게임을 실행한다.

as-is:
public class RacingGame {
    private static final Integer CAR_NAME_MAX_LENGTH = 5;
    private static final String NUMBER_PATTERN = "^[0-9]*$";
    pirvate final ArryaList<Car> cars;
    private int turns;
    public RacingGame() {...}
    public RacingGame(ArryaList<Car> cars) {...}
    private void inputRacingGame() throws IllegalArgumentException {...}
    private void inputRacingTurns() throws IllegalArgumentException {...}
    private void inputRacingInformation() {...}
    private void printThisTurnResult() {...}
    private void startNextTurn() {...}
    private void printWinnerList() {...}
    public void startGame() {...}
}

to-be:
public class RacingGame {
    private static final Integer CAR_NAME_MAX_LENGTH = 5;
    private static final String NUMBER_PATTERN = "^[0-9]*$";
    private final Race race;
    public RacingGame() {...}
    public RacingGame(Race race) {...}
    private void inputRacingCarName() throws IllegalArgumentException
    {...}
    private void inputRacingTurns() throws IllegalArgumentException {...}
    private void inputRacingInformation() {...}
    public void startRacingGame() {...}
}

public class Race {
    private final ArryaList<Car> cars;
    private int turns;
    private int farthestPosition;
    public Race() {...}
    public Race(final Array<Car> cars) {...}
    private void printThisTurnResult() {...}
    private void startNextTurn() {...}
    private void printWinnerList() {...}
    pulbic void startRace() {...}
    public void addCar(final Car car) {...}
    public void removeCarList() {...}
    public void setTurns(final int turns) {...}
}
